fix: renamed modules to align with pep8 recommendations (added underscores to improve readability)
