+++ app/modules/all_descriptors.py
